refactor: remove streamer.QueueStats
This updates qlast to use the queuestreamer impl.
