--- pkg/be/events/qstat/options.go
@@ -1,8 +1,0 @@
-package qstat
-
-type Options struct {
-	Follow  bool
-	Tail    int64
-	Verbose bool
-	Quiet   bool
-}
