--- pkg/be/ibmcloud/qstat.go
@@ -1,11 +1,0 @@
-package ibmcloud
-
-import (
-	"fmt"
-
-	"lunchpail.io/pkg/be/events/qstat"
-)
-
-func (streamer Streamer) QueueStats(c chan qstat.Model, opts qstat.Options) error {
-	return fmt.Errorf("Unsupported operation: 'StreamQueueStats'")
-}
